clean up and move read_proj_info function from dmft to embed
