+++ src/python/__init__.py
@@ -47,7 +47,7 @@ from .utils import set_verbosity, MpiHandler
 from .mean_field import make_mf
 from .interaction import make_thc_coulomb, make_chol_coulomb, run_isdf
 from .mbpt import run_hf, run_gw, run_qpg0w0
-from .embed import downfold_1e, downfold_2e, downfold_local_gf, downfold_coulomb, dmft_embed
+from .embed import read_proj_info, downfold_1e, downfold_2e, downfold_local_gf, downfold_coulomb, dmft_embed
 from .wannier import wannier90, coqui2wannier90, append_wannier90_win
 
 # Make submodules available
