+++ src/python/embed/__init__.py
@@ -18,6 +18,6 @@ limitations under the License.
 ==========================================================================
 """
 
-from .embed import downfold_1e, downfold_2e, downfold_local_gf, downfold_coulomb, dmft_embed
+from .embed import read_proj_info, downfold_1e, downfold_2e, downfold_local_gf, downfold_coulomb, dmft_embed
 
-__all__ = ["downfold_1e", "downfold_2e", "downfold_local_gf", "downfold_coulomb", "dmft_embed"]
+__all__ = ["downfold_1e", "downfold_2e", "downfold_local_gf", "downfold_coulomb", "dmft_embed", "read_proj_info"]
