+++ src/python/embed/tests/test_downfold.py
@@ -47,7 +47,7 @@ def test_downfold(mpi):
     coqui.run_gw(gw_params, h_int=thc)
     mpi.barrier()
 
-    #proj_info = read_proj_info(mf.outdir()+"/lih_wan.h5")
+    proj_info = coqui.read_proj_info(mf.outdir()+"/lih_wan.h5")
 
     # downfold the local Green's function
     gloc_params = {
@@ -56,7 +56,7 @@ def test_downfold(mpi):
         "wannier_file": mf.outdir() + "/lih_wan.h5",
         "force_real": True
     }
-    Gloc_t = coqui.downfold_local_gf(mf, gloc_params)#, projector_info=proj_info)
+    Gloc_t = coqui.downfold_local_gf(mf, gloc_params)
 
     assert np.allclose(Gloc_t.imag, 0.0, atol=1e-10), "Imaginary part of Gloc(t) is not negligible"
     assert Gloc_t[-1,0,0,0] == pytest.approx(-0.9921589287308954, abs=1e-10)
@@ -71,7 +71,7 @@ def test_downfold(mpi):
         "wannier_file": mf.outdir() + "/lih_wan.h5",
         "permut_symm": True, "force_real": True, "output_in_tau": True
     }
-    Vloc, Wloc_t = coqui.downfold_coulomb(thc, wloc_params)#, projector_info=proj_info)
+    Vloc, Wloc_t = coqui.downfold_coulomb(thc, wloc_params)
 
     assert np.allclose(Vloc.imag, 0.0, atol=1e-12), "Imaginary part of Vloc is not negligible"
     assert Vloc[0,0,0,0] == pytest.approx(1.4160723518754155, abs=1e-12)
@@ -87,7 +87,7 @@ def test_downfold(mpi):
 
     # downfold the cRPA local screened interaction
     wloc_params["screen_type"] = "crpa"
-    Vloc, Uloc_t = coqui.downfold_coulomb(thc, wloc_params)#, projector_info=proj_info)
+    Vloc, Uloc_t = coqui.downfold_coulomb(thc, wloc_params, projector_info=proj_info)
 
     assert np.allclose(Uloc_t.imag, 0.0, atol=1e-12), "Imaginary part of Uloc(t) is not negligible"
     assert Uloc_t[0,0,0,0,0] == pytest.approx(-0.2152982864315092, abs=1e-12)
@@ -118,7 +118,7 @@ def test_local_coulomb_from_mf(mpi):
         "wannier_file": mf.outdir() + "/lih_wan.h5",
         "permut_symm": True, "force_real": True, "output_in_tau": True
     }
-    Vloc, Wloc_t = coqui.downfold_coulomb(thc, wloc_params)#, projector_info=proj_info)
+    Vloc, Wloc_t = coqui.downfold_coulomb(thc, wloc_params)
 
     assert np.allclose(Vloc.imag, 0.0, atol=1e-12), "Imaginary part of Vloc is not negligible"
     assert Vloc[0,0,0,0] == pytest.approx(1.4160723518754155, abs=1e-12)
